refactor: single source of truth for reserved keywords per language
Reserved keyword sets were duplicated across the IR generators, the
IrEmitter companions and the Avro extension. They now live once per
language as Keywords objects in ir/generator/ReservedKeywords.kt:

- generators escape identifiers against the shared objects
- emitter companions delegate via `Keywords by XxxKeywords`, keeping
  the existing API for consumers (e.g. jackson)
- AvroExtension drops its inlined Java/Kotlin copies

The per-language sets are the union of the previous copies. RustGenerator
keeps a documented exception: `self`/`Self` cannot be raw-escaped
(r#self is invalid Rust), so they are excluded from generator escaping
and handled by the emitter's field renaming instead.

The Wirespec emitter's keyword set stays in WirespecIdentifierEmitter -
its module does not depend on the IR module, and it is already the only
definition for that language.
