+++ src/compiler/emitters/python/src/commonMain/kotlin/community/flock/wirespec/emitters/python/PythonIrEmitter.kt
@@ -39,6 +39,7 @@ import community.flock.wirespec.ir.core.raw
 import community.flock.wirespec.ir.core.transform
 import community.flock.wirespec.ir.core.transformChildren
 import community.flock.wirespec.ir.emit.IrEmitter
+import community.flock.wirespec.ir.generator.PythonKeywords
 import community.flock.wirespec.ir.emit.placeInModule
 import community.flock.wirespec.ir.emit.prependImports
 import community.flock.wirespec.ir.generator.PythonGenerator
@@ -308,16 +309,6 @@ open class PythonIrEmitter(
         import(wirespecPath, "_raise"),
     )
 
-    companion object : Keywords {
-        override val reservedKeywords = setOf(
-            "False", "None", "True", "and", "as", "assert",
-            "break", "class", "continue", "def", "del",
-            "elif", "else", "except", "finally", "for",
-            "from", "global", "if", "import", "in",
-            "is", "lambda", "nonlocal", "not", "or",
-            "pass", "raise", "return", "try", "while",
-            "with", "yield"
-        )
-    }
+    companion object : Keywords by PythonKeywords
 
 }
