+++ src/compiler/ir/src/commonMain/kotlin/community/flock/wirespec/ir/generator/JavaGenerator.kt
@@ -637,18 +637,4 @@ object JavaGenerator : Generator {
     )
 }
 
-private fun String.sanitize(): String = if (reservedKeywords.contains(this)) "_$this" else this
-
-private val reservedKeywords = setOf(
-    "abstract", "continue", "for", "new", "switch",
-    "assert", "default", "if", "package", "synchronized",
-    "boolean", "do", "goto", "private", "this",
-    "break", "double", "implements", "protected", "throw",
-    "byte", "else", "import", "public", "throws",
-    "case", "enum", "instanceof", "return", "transient",
-    "catch", "extends", "int", "short", "try",
-    "char", "final", "interface", "static", "void",
-    "class", "finally", "long", "strictfp", "volatile",
-    "const", "float", "native", "super", "while",
-    "true", "false", "null",
-)
+private fun String.sanitize(): String = if (JavaKeywords.reservedKeywords.contains(this)) "_$this" else this
