+++ src/compiler/ir/src/commonMain/kotlin/community/flock/wirespec/ir/generator/ScalaGenerator.kt
@@ -633,16 +633,4 @@ object ScalaGenerator : Generator {
     private fun TypeDescriptor.emitTypeDescriptor(): String = "scala.reflect.classTag[${type.emitGenerics()}]"
 }
 
-private fun String.sanitize(): String = if (reservedKeywords.contains(this)) "`$this`" else this
-
-private val reservedKeywords = setOf(
-    "abstract", "case", "class", "def", "do",
-    "else", "extends", "false", "final", "for",
-    "forSome", "if", "implicit", "import", "lazy",
-    "match", "new", "null", "object", "override",
-    "package", "private", "protected", "return", "sealed",
-    "super", "this", "throw", "trait", "true",
-    "try", "type", "val", "var", "while",
-    "with", "yield", "given", "using", "enum",
-    "export", "then",
-)
+private fun String.sanitize(): String = if (ScalaKeywords.reservedKeywords.contains(this)) "`$this`" else this
